+++ src/main/java/org/ballerina/deployment/FrameworkConstants.java
@@ -15,11 +15,15 @@
 *  specific language governing permissions and limitations
 *  under the License.
 */
-package org.wso2.ballerina.deployment.Base;
+package org.ballerina.deployment;
 
-public class BallerinaConstants {
-
-    public static String ENABLE_URL_REMAP = "ENABLE_REMAP";
-    public static String POD_TAG_NAME = "ballerina";
+/**
+ * FrameworkConstants.
+ */
+public class FrameworkConstants {
+    public static final String SYSTEM_ARTIFACT_RESOURCE_LOCATION = "framework.resource.location";
+    public static final String JSON_FILE_PATH = "jsonFilePath";
+    public static final String DEFAULT_TESTNG_FILE = "testng.xml";
+    public static final String TESTNG_RESULT_OUT_DIRECTORY = "intergration-tests";
 
 }
