+++ src/main/java/org/ballerina/deployment/beans/Deployment.java
@@ -15,7 +15,7 @@
 *  specific language governing permissions and limitations
 *  under the License.
 */
-package org.wso2.ballerina.deployment.beans;
+package org.ballerina.deployment.beans;
 
 import java.util.HashMap;
