+++ src/main/java/org/ballerina/deployment/commons/DeploymentYamlConstants.java
@@ -15,7 +15,7 @@
 *  specific language governing permissions and limitations
 *  under the License.
 */
-package org.wso2.ballerina.deployment.commons;
+package org.ballerina.deployment.commons;
 
 /**
  * Deployment Yaml Constants.
