+++ src/test/resources/Test-suites/testng_pattern1.xml
@@ -20,7 +20,7 @@
 <suite name="ballerina_standalone">
     <test name="ballerina-tests" preserve-order="true" parallel="false">
         <classes>
-            <class name="org.wso2.ballerina.deployment.routing.EchoService"/>
+            <class name="org.ballerina.tests.routing.EchoService"/>
         </classes>
     </test>
 </suite>
